bench/requirements.txt
locust
locust-plugins
websockets
